rename PauliNet to NeuralNetworkWaveFunction
