+++ src/deepqmc/wf/__init__.py
@@ -1,4 +1,4 @@
 from .base import WaveFunction
-from .paulinet import PauliNet
+from .nn_wave_function import NeuralNetworkWaveFunction
 
-__all__ = ['PauliNet', 'WaveFunction']
+__all__ = ['NeuralNetworkWaveFunction', 'WaveFunction']
