+++ src/deepqmc/wf/nn_wave_function/__init__.py
@@ -1,0 +1,3 @@
+from .nn_wave_function import NeuralNetworkWaveFunction
+
+__all__ = ['NeuralNetworkWaveFunction']
