+++ src/deepqmc/wf/nn_wave_function/cusp.py
