+++ src/deepqmc/wf/nn_wave_function/env.py
