+++ src/deepqmc/wf/nn_wave_function/omni.py
