--- src/deepqmc/wf/paulinet/__init__.py
@@ -1,3 +1,0 @@
-from .paulinet import PauliNet
-
-__all__ = ['PauliNet']
